tests/requirements.txt
atoma==0.0.17
pytest==5.3.0
pytest-django==3.7.0
selenium==3.141.0
pytest-mock==1.12.1